Rewrite crate to enable reusing of (response)`Id`s
The `Id` (or `AwaitableResponse`) is actually an encapsulate of
`response_id`.

Since each request is identified by a unique id and the read_end needs
to quickly look up one, it is stored in `Arena` for quick lookup,
insertion and removal.

However, since `Arena` is thread safe, insertion and removal has
significant overhead, mostly because of synchronoisation.

Reusing them reduce this overhead by avoiding the insertion and removal
all together.

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>