Support setting time in MetaDataBuilder (#124)
The time fields (accessed, modified) in MetaData are currently not changeable, but it's supported by the protocol.

Setting time attributes of a file could be useful for many cases, especially to workaround the coarse timestamp granularity (seconds) of sftp. By setting timestamps, we can then tell if a file has been modified within a second.